refactored book category component of Home page
