+++ src/components/bookCardComponent/BookCard.module.css
@@ -3,7 +3,8 @@
     z-index: 1;
     display: block;
     background: #FFFFFF;
-    height: 340px;
+    width: 31.366rem;
+    height: 28.545rem;
     overflow: hidden;
     -webkit-box-shadow: 0px 1px 2px 0px rgba(0, 0, 0, 0.15);
     -moz-box-shadow: 0px 1px 2px 0px rgba(0, 0, 0, 0.15);
@@ -13,6 +14,7 @@
     -ms-transition: all 0.3s linear 0s;
     -o-transition: all 0.3s linear 0s;
     transition: all 0.3s linear 0s;
+    border: 2px solid yellow;
 }
 
 .bookCard:hover,
@@ -34,7 +36,7 @@
 
 .bookCard img {
     display: block;
-    width: 120%;
+    width: 100%;
     -webkit-transition: all 0.3s linear 0s;
     -moz-transition: all 0.3s linear 0s;
     -ms-transition: all 0.3s linear 0s;
