+++ src/components/container/Container.module.css
@@ -7,6 +7,7 @@
     margin: 0 auto;
     justify-content: space-between;
     align-items: center;
+    border: 2px solid red;
 }
 
 @media screen and (max-width: 768px) {
